extracted derived field functions into common trait
