+++ src/utils/mod.rs
@@ -1,4 +1,4 @@
-pub fn uninit_vector(length: usize) -> Vec<u64> {
+pub fn uninit_vector<T>(length: usize) -> Vec<T> {
     let mut vector = Vec::with_capacity(length);
     unsafe { vector.set_len(length); }
     return vector;
